Incluindo usuario no banco de dados e arrumando imports e testes unitarios

+++ products-api/app/database/products_dao.py
@@ -1,24 +1,27 @@
+import logging
+
 import pymysql
 from pymysql import OperationalError
 
-from app.model.product import Product
 from app import environment_config
-from exceptions import exceptions
-import logging
+from app.model.product import Product
+from exceptions import exceptions, ApiException
 
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.DEBUG)
 
-SQL_QUERY_SELECT_ALL_PRODUCTS = "SELECT ID, NAME, PRICE, URL FROM TB_PRODUCT"
-SQL_QUERY_SELECT_PRODUCT = "SELECT ID, NAME, PRICE, URL FROM TB_PRODUCT WHERE ID = %s"
-SQL_QUERY_INSERT_PRODUCT = "INSERT INTO TB_PRODUCT (ID, NAME, PRICE, URL) VALUES (%s, %s, %s, %s)"
-SQL_QUERY_UPDATE_PRODUCT = "UPDATE TB_PRODUCT SET NAME=%s, PRICE=%s, URL=%s WHERE ID = %s"
-SQL_QUERY_DELETE_PRODUCT = "DELETE FROM TB_PRODUCT WHERE ID = %s"
+SQL_QUERY_SELECT_ALL_PRODUCTS = "SELECT ID, NAME, PRICE, URL FROM TB_PRODUCT WHERE USER_ID = %s"
+SQL_QUERY_SELECT_PRODUCT = "SELECT ID, NAME, PRICE, URL FROM TB_PRODUCT WHERE ID = %s AND USER_ID = %s"
+SQL_QUERY_INSERT_PRODUCT = "INSERT INTO TB_PRODUCT (ID, NAME, PRICE, URL, USER_ID) VALUES (%s, %s, %s, %s, %s)"
+SQL_QUERY_UPDATE_PRODUCT = "UPDATE TB_PRODUCT SET NAME=%s, PRICE=%s, URL=%s WHERE ID = %s AND USER_ID = %s"
+SQL_QUERY_DELETE_PRODUCT = "DELETE FROM TB_PRODUCT WHERE ID = %s AND USER_ID = %s"
 
 SQL_QUERY_SELECT_ALL_CATEGORIES_BY_PRODUCT = "SELECT NAME, PRODUCT_ID FROM TB_CATEGORY WHERE PRODUCT_ID = %s"
 SQL_QUERY_INSERT_CATEGORY = "INSERT INTO TB_CATEGORY (NAME, PRODUCT_ID) VALUES (%s, %s)"
 SQL_QUERY_DELETE_CATEGORY_BY_PRODUCT = "DELETE FROM TB_CATEGORY WHERE PRODUCT_ID = %s"
 
 
-def find_all() -> []:
+def find_all(auth_id: str) -> []:
     conn = None
     try:
         conn = pymysql.connect(**environment_config.get_db_config())
@@ -26,7 +29,7 @@ def find_all() -> []:
 
         with conn:
             cursor = conn.cursor()
-            cursor.execute(SQL_QUERY_SELECT_ALL_PRODUCTS)
+            cursor.execute(SQL_QUERY_SELECT_ALL_PRODUCTS, auth_id)
 
             product_rows = cursor.fetchall()
             for product_row in product_rows:
@@ -49,19 +52,16 @@ def find_all() -> []:
         return products
 
     except OperationalError as error:
-        logging.error("ERROR: Unexpected error: Could not connect to MySql instance: {}".format(error))
+        logger.error("ERROR: Unexpected error: Could not connect to MySql instance: {}".format(error))
         if conn:
             conn.rollback()
         raise error
-
-    except BaseException as error:
-        logging.error("ERROR: Unexpected error: {}".format(error))
+    except ApiException as error:
         if conn:
             conn.rollback()
         raise error
-
     except Exception as error:
-        logging.error("ERROR: Unexpected error: {}".format(error))
+        logger.error("ERROR: Unexpected error: {}".format(error))
         if conn:
             conn.rollback()
         raise error
@@ -70,14 +70,14 @@ def find_all() -> []:
             conn.close()
 
 
-def find_one(product_id: str) -> Product:
+def find_one(auth_id: str, product_id: str) -> Product:
     conn = None
     try:
         conn = pymysql.connect(**environment_config.get_db_config())
 
         with conn:
             cursor = conn.cursor()
-            cursor.execute(SQL_QUERY_SELECT_PRODUCT, product_id)
+            cursor.execute(SQL_QUERY_SELECT_PRODUCT, (product_id, auth_id))
 
             result = cursor.fetchone()
 
@@ -103,19 +103,16 @@ def find_one(product_id: str) -> Product:
             return product
 
     except OperationalError as error:
-        logging.error("ERROR: Unexpected error: Could not connect to MySql instance: {}".format(error))
+        logger.error("ERROR: Unexpected error: Could not connect to MySql instance: {}".format(error))
         if conn:
             conn.rollback()
         raise error
-
-    except BaseException as error:
-        logging.error("ERROR: Unexpected error: {}".format(error))
+    except ApiException as error:
         if conn:
             conn.rollback()
         raise error
-
     except Exception as error:
-        logging.error("ERROR: Unexpected error: {}".format(error))
+        logger.error("ERROR: Unexpected error: {}".format(error))
         if conn:
             conn.rollback()
         raise error
@@ -133,7 +130,8 @@ def insert(product: Product):
 
         with conn:
             cursor = conn.cursor()
-            cursor.execute(SQL_QUERY_INSERT_PRODUCT, (product.id, product.name, product.price, product.url))
+            cursor.execute(SQL_QUERY_INSERT_PRODUCT,
+                           (product.id, product.name, product.price, product.url, product.user_id))
 
             if cursor.rowcount == 0:
                 exception = exceptions.get(exceptions.PRODUCT_SAVE_TO_DATABASE_EXCEPTION)
@@ -154,19 +152,16 @@ def insert(product: Product):
             conn.commit()
 
     except OperationalError as error:
-        logging.error("ERROR: Unexpected error: Could not connect to MySql instance: {}".format(error))
+        logger.error("ERROR: Unexpected error: Could not connect to MySql instance: {}".format(error))
         if conn:
             conn.rollback()
         raise error
-
-    except BaseException as error:
-        logging.error("ERROR: Unexpected error: {}".format(error))
+    except ApiException as error:
         if conn:
             conn.rollback()
         raise error
-
     except Exception as error:
-        logging.error("ERROR: Unexpected error: {}".format(error))
+        logger.error("ERROR: Unexpected error: {}".format(error))
         if conn:
             conn.rollback()
         raise error
@@ -184,7 +179,7 @@ def update(product: Product):
 
         with conn:
             cursor = conn.cursor()
-            cursor.execute(SQL_QUERY_UPDATE_PRODUCT, (product.name, product.price, product.url, product.id))
+            cursor.execute(SQL_QUERY_UPDATE_PRODUCT, (product.name, product.price, product.url, product.id, product.user_id))
 
             cursor.execute(SQL_QUERY_DELETE_CATEGORY_BY_PRODUCT, product.id)
             for category in product.categories:
@@ -200,19 +195,16 @@ def update(product: Product):
             conn.commit()
 
     except OperationalError as error:
-        logging.error("ERROR: Unexpected error: Could not connect to MySql instance: {}".format(error))
+        logger.error("ERROR: Unexpected error: Could not connect to MySql instance: {}".format(error))
         if conn:
             conn.rollback()
         raise error
-
-    except BaseException as error:
-        logging.error("ERROR: Unexpected error: {}".format(error))
+    except ApiException as error:
         if conn:
             conn.rollback()
         raise error
-
     except Exception as error:
-        logging.error("ERROR: Unexpected error: {}".format(error))
+        logger.error("ERROR: Unexpected error: {}".format(error))
         if conn:
             conn.rollback()
         raise error
@@ -221,7 +213,7 @@ def update(product: Product):
             conn.close()
 
 
-def delete(product_id: str):
+def delete(auth_id: str, product_id: str):
     conn = None
     try:
 
@@ -230,24 +222,29 @@ def delete(product_id: str):
 
         with conn:
             cursor = conn.cursor()
+
+            cursor.execute(SQL_QUERY_SELECT_PRODUCT, (product_id, auth_id))
+            cursor.fetchone()
+            if cursor.rowcount == 0:
+                exception = exceptions.get(exceptions.PRODUCT_NOT_FOUND_EXCEPTION)
+                exception.append_to_log("Couldn't find product: {}".format(product_id))
+                raise exception
+
             cursor.execute(SQL_QUERY_DELETE_CATEGORY_BY_PRODUCT, product_id)
-            cursor.execute(SQL_QUERY_DELETE_PRODUCT, product_id)
+            cursor.execute(SQL_QUERY_DELETE_PRODUCT, (product_id, auth_id))
             conn.commit()
 
     except OperationalError as error:
-        logging.error("ERROR: Unexpected error: Could not connect to MySql instance: {}".format(error))
+        logger.error("ERROR: Unexpected error: Could not connect to MySql instance: {}".format(error))
         if conn:
             conn.rollback()
         raise error
-
-    except BaseException as error:
-        logging.error("ERROR: Unexpected error: {}".format(error))
+    except ApiException as error:
         if conn:
             conn.rollback()
         raise error
-
     except Exception as error:
-        logging.error("ERROR: Unexpected error: {}".format(error))
+        logger.error("ERROR: Unexpected error: {}".format(error))
         if conn:
             conn.rollback()
         raise error
